+++ tests/functional_test.py
@@ -35,7 +35,7 @@ def test_one_hot_invalid():
 
     msg = "OneHot num classes must be specified and must be constant."
     # Run on IPU.
-    with pytest.raises(RuntimeError, match=msg):
+    with pytest.raises(poptorch.Error, match=msg):
         poptorch_model = poptorch.inferenceModel(model)
         poptorch_model(input)